refactor: UI관련 불필요한 파일 삭제 및 DesignSystem폴더로 이동
