--- HongikYeolgong2-iOS/General/Extension/UIColor+.swift
@@ -1,110 +1,0 @@
-
-import UIKit
-import Foundation
-
-extension UIColor {
-    
-    //hexcode: D3D6E0
-    static let customGray100 = UIColor(
-        red: 0.827,
-        green: 0.839,
-        blue: 0.878,
-        alpha: 1.0
-    )
-    
-    //hexcode: A5A8B0
-    static let customGray200 = UIColor(
-        red: 0.647,
-        green: 0.659,
-        blue: 0.69,
-        alpha: 1.0
-    )
-    
-    //hexcode: 65676B
-    static let customGray300 = UIColor(
-        red: 0.396,
-        green: 0.404,
-        blue: 0.42,
-        alpha: 1.0
-    )
-    
-    //hexcode: 4F515A
-    static let customGray400 = UIColor(
-        red: 0.31,
-        green: 0.318,
-        blue: 0.353,
-        alpha: 1.0
-    )
-    
-    //hexcode: 3A3E49
-    static let customGray600 = UIColor(
-        red: 0.227,
-        green: 0.243,
-        blue: 0.286,
-        alpha: 1.0
-    )
-    
-    //hexcode: 23262D
-    static let customGray800 = UIColor(
-        red: 0.137,
-        green: 0.149,
-        blue: 0.176,
-        alpha: 1.0
-    )
-    
-    //hexcode: 4164E2
-    static let customBlue100 = UIColor(
-        red: 0.255,
-        green: 0.392,
-        blue: 0.886,
-        alpha: 1.0
-    )
-    
-    //hexcode: 2747BA
-    static let customBlue200 = UIColor(
-        red: 0.153,
-        green: 0.278,
-        blue: 0.729,
-        alpha: 1.0
-    )
-    
-    //hexcode: 213065
-    static let customBlue300 = UIColor(
-        red: 0.129,
-        green: 0.188,
-        blue: 0.396,
-        alpha: 1.0
-    )
-    
-    //hexcode: 121C42
-    static let customBlue400 = UIColor(
-        red: 0.071,
-        green: 0.11,
-        blue: 0.259,
-        alpha: 1.0
-    )
-    
-    //hexcode: F4FD3D
-    static let customYellow100 = UIColor(
-        red: 0.957,
-        green: 0.992,
-        blue: 0.239,
-        alpha: 1.0
-    )
-    
-    //hexcode: 4164E2
-    static let customYellow300 = UIColor(
-        red: 0.808,
-        green: 0.812,
-        blue: 0.216,
-        alpha: 1.0
-    )
-    
-    //hexcode: 0C0D11
-    static let customBackground = UIColor(
-        red: 0.047,
-        green: 0.050,
-        blue: 0.066,
-        alpha: 1.0
-    )
-}
